added route and initial react based component to create an event, and restructured some files inside calendar module
