+++ public/calendar/event.edit/event.edit.html
@@ -1,0 +1,1 @@
+<event-edit></event-edit>
