+++ public/calendar/react.big.calendar/calendar.jsx
@@ -28,7 +28,7 @@ class CbsFullCalendar extends React.Component {
 			<BrowserRouter>
 				<div className='h-80vh'>
 					<div>
-						<Link className='btn' to="/calendar/new">Add New Event</Link>
+						<Link className='btn' to="/calendar/edit/event">Add New Event</Link>
 					</div>
 					<BigCalendar
 						localizer = {localizer}
